@@ -1,6 +1,6 @@
 {
   "name": "react-keyboard-navigation-popup",
-  "version": "1.0.0",
+  "version": "1.0.1",
   "description": "wip",
   "author": "aleqsio",
   "license": "MIT",
@@ -47,7 +47,7 @@
     "rollup-plugin-postcss": "^1.6.2",
     "rollup-plugin-typescript2": "^0.17.0",
     "rollup-plugin-url": "^1.4.0",
-    "typescript": "^2.8.3"
+    "typescript": "^4.1.3"
   },
   "files": [
     "dist"
